refactor: introduce BaseHub<T> to centralize connection management for SignalR hubs
Abstracted common OnConnectedAsync and OnDisconnectedAsync logic into a generic BaseHub<T> class to eliminate code duplication across hubs such as SignalingHub and NotificationsHub.
