+++ Interfaces/ISignalingClient.cs
@@ -1,0 +1,8 @@
+namespace AonFreelancing.Interfaces;
+
+public interface ISignalingClient
+{
+    public Task GetOffer(string offerJson);
+    public Task GetAnswer(string answerJson);
+    public Task GetIceCandidate(string iceCandidateJson);
+}
